# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  version: "1.0.0"
...
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authorization header using the Bearer
      scheme: bearer
      bearerFormat: JWT    
  schemas:
        Id:
          type: string
          description: The unique identifier for the user.
          readOnly: True
          pattern: "{09az}{8}-{09az}{4}-{09az}{4}-{09az}{4}-{09az}{12}"
          minLength: 36
          maxLength: 36
        Username:
          type: string
          description: The username of the user.
          example: "mario"
          readOnly: True
          pattern: '^.*$'
          minLength: 3
          maxLength: 20
...
        Date:
          description: A date represented by year, month, and day components(YYYY-MM-DD).
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          minLength: 10
          maxLength: 10
        Text:
          description: The text of the comment.
          type: string
          pattern: '^.*?$'
          minLength: 0
          maxLength: 300
...
        User:
          title: User
          description: User registered to WASAPhoto
          type: object
          properties: 
            userId:
              $ref: "#/components/schemas/Id"
            username:
              $ref: "#/components/schemas/Username"
          required:
          - userId
          - username
...
        Like:
          title: Like
          type: object
          description: Like of a photo
          properties:
            likeId:
              $ref: "#/components/schemas/Id"
            photoId:
              $ref: "#/components/schemas/Id"
            author:
              $ref: "#/components/schemas/Id"
          required:
            - likeId
            - photoId
            - author
        Comment:
          title: Comment
          description: Comment of a photo
          type: object
          properties:
            photoId:
              $ref: "#/components/schemas/Id"
            commentId:
              $ref: "#/components/schemas/Id"
            author:
              $ref: "#/components/schemas/Id"
            text:
              $ref: "#/components/schemas/Text"
            date:
                $ref: "#/components/schemas/Date"
          required:
            - photoId
            - commentId
            - author
            - text
            - date
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: ["login"]
      summary: Log in the user
      description: |
        If the use does not exist, it will be created,
        and an identifier is returned.
        If the user already exists, the identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Username" }
        required: true

      responses:
        '201':
          description: User created
          content:
            application/json:  
              schema: { $ref: "#/components/schemas/User" }

        '200': # return the user
          description: User already exists 
          content:
            application/json:  
              schema: { $ref: "#/components/schemas/User" }
...
  /photos/{photoId}: #deletePhoto
    delete:
      tags: ["photo"]
      operationId: deletePhoto
      summary: delete a photo
      description: |
        Delete a photo.
        The user must be already logged in.
      parameters:
        - name: photoId
          in: path
          description: The identifier of the photo
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        - name: userId
          in: query
          description: The identifier of the user used as a form of authentication.
          required: true
          schema: { $ref: "#/components/schemas/Id"}
      responses:
        "204":
          description: "photo succesfully deleted"
        "403":
          description: |
            User not authorized to delete a photo, not the owner of the photo
        "404":
          description: photo not found 
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /follow/: # followUser
    post:
      tags: ["following"]
      summary: Follow a user
      description: |
        Follow a user.
        The user must be already logged in.
      operationId: followUser
      parameters: 
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
      requestBody:
        description: User to follow.
        content:
         application/json:
            schema:
              type: object
              description: User to follow
              properties:
                userId:
                  $ref: "#/components/schemas/Id"
                username:
                  $ref: "#/components/schemas/Username"
        required: true
      responses:
        "201":
          description: User succesfully followed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" } 
        "403":
          description: |
            User not authorized to follow himself 
            or a user already followed
            or user blocked by the user tried to follow
        "404":
         description: User to follow not found
        "503":
          description: |
            Too many requests to the server requestin to follow that user
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /ban/: #banUser
    post:
      tags: ["ban"]
      operationId: banUser
      summary: Ban a user
      description: |
        Ban a user.
        The user must be already logged in.
      parameters:
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
      requestBody:
        description: Username to ban and userId of the user who is bannig.
        content:
          application/json:
            schema:
              type: object
              description: Username to ban and userId of the user who is banning
              properties:
                username:
                  $ref: "#/components/schemas/Username"
                userId:
                  $ref: "#/components/schemas/Id"

        required: true 
      responses:
        "201":
          description: User succesfully added to the ban list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: |
            User not authorized to ban himself 
            or a user already banned
            or user blocked by the user tried to ban
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /users/{username}: # getUserProfile & setMyUserName
      parameters:
          - name: username
            in: path
            description: The username of the user.
            required: true
            schema: { $ref: "#/components/schemas/Username"} 
      get:
        tags: ["user"]
        operationId: getUserProfile
        summary: Get the profile of a user
        description: |
          Get the profile of a user.
          The user must be already logged in. 
        parameters: 
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        responses:
          "200":
            description: User succesfully found
            content:
              application/json:
                schema: { $ref: "#/components/schemas/User" }
          "404":
            description: User not found, maybe deleted the account
          "403":
            description: |
             User banned by the user tried to get the profile
          "401":
            $ref: "#/components/responses/UnauthorizedError"

      put: 
        tags: ["user"]
        operationId: setMyUserName
        summary: Set the username of the user
        description: |
          Set the username of the user.
          The username must be unique.
          The user must be already logged in.   
        parameters: 
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        - name: userId
          in: query
          description: The identifier of the user used as a form of authentication.
          required: true
          schema: { $ref: "#/components/schemas/Id"}
        requestBody:
          description: New username
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Username" }
          required: true        
        responses:
          "200":
            description: Username succesfully set
            content:
              application/json:
                schema: { $ref: "#/components/schemas/User" }
          "404":
            description: User don't exist
          "401":
            $ref: "#/components/responses/UnauthorizedError"
...
  /likes/: #likePhoto 
    post:
      tags: ["like"]
      summary: Like a photo
      description: |
        Like a photo.
        The user must be already logged in.
      operationId: likePhoto
      parameters:
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
      requestBody:
        description: Photo and description
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 2
              maxItems: 2
              description: Photo and author
              items:
                type: object
                description: photoId and authorId
                properties:
                  photoId:
                    $ref: "#/components/schemas/Id"
                  authorId:
                    $ref: "#/components/schemas/Id"
      responses:
        "201":
          description: Photo succesfully liked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Like" }
        "403":
          description: |
            User not authorized to like a photo, blocked by the owner 
            of the photo (photo retrieved by a link or by other form of sharing)

        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /comments/: #commentPhoto
    post:
      tags: ["comment"]
      summary: Comment a photo
      description: |
        Comment a photo.
        The user must be already logged in.
      operationId: commentPhoto
      parameters:
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
      requestBody:
        description: Photo and description
        required: true
        content:
          multipart/form-data:
            schema:
              type: array
              minItems: 3
              maxItems: 3
              description: Photo, author and text
              items:
                type: object
                description: photoId, authorId and text
                properties:
                  photoId:
                    $ref: "#/components/schemas/Id"
                  authorId:
                    $ref: "#/components/schemas/Id"
                  text:
                    $ref: "#/components/schemas/Text"
      responses:
        "201":
          description: Photo succesfully commented  
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" } 
        "404":
          description: Photo not found
        "403":
          description: |
            Photo deleted, blocked by the owener of the photo (photo retrieved by a link or by other form of sharing)
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...